feat: finalize EspNow manager coverage and introduce IEspNow interface
- Achieved near 100% test coverage for espnow_manager.cpp.
- Introduced IEspNow interface in include/espnow_interfaces.hpp for better decoupling.
- Moved EspNowConfig to include/espnow_types.hpp to resolve circular dependencies.
- Added test cases for WiFi channel switching logic (Heartbeat/Scan responses).
- Added test cases for stored peer re-registration during initialization.
- Added test cases for send_command and static ESP-NOW callbacks.
- Verified all 12 host test suites pass successfully.
